pyproject.toml
[project]
name = "appointment-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch<9.0,>=8.0",
    "fastembed>=0.7.1",
    "ipykernel>=6.29.5",
    "minsearch>=0.0.3",
    "notebook==7.1.2",
    "openai>=1.90.0",
    "pandas>=2.3.0",
    "poetry>=2.1.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sentence-transformers==2.7.0",
    "tdqm>=0.0.1",
    "qdrant-client[fastembed]>=1.14.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "google-generativeai>=0.8.5",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
